Really close all the executors when the bean is destroyed - Issue #1094
Previously we were doing this in a @PreDestroy method on the
interceptor. However, that method will only fire if the interceptor is
bound to a class. It won't fire if the interceptor is just bound to a
method.

Since we regularly bind the FT Interceptor to methods, that's not
acceptable. The new solution injects a dependent bean into the
interceptor and the cleanup is done in its PreDestroy method.