Make messages which reference a method less verbose
Previously these messages were putting in java.lang.Method.toString()
which is very verbose and includes all information about the method.

In an error message, this is not useful. We just want enough
information for the user to identify the problematic method.

Also added the bndtools nature to
com.ibm.ws.microprofile.faulttolerance.spi because it was missing.