Increase core threads for FaultTolerance tests
Timeout tests are sensitive to timing. If a test fills the thread pool,
it takes a second or so for the thread pool to notice the increase in
load and add more threads.

By setting a larger number of core threads, we avoid this unpredictable
delay and eliminate a difference in behaviour between test machines with
different numbers of hardware cores which should result in tests passing
or failing more reliably.

Also added light ThreadPoolController trace to help debug timeout tests
in the future.

For issue #214