Return true if no classes are in the modified file list
The redefineClasses method contract states that it will return true when
the passed-in Notification contains no class files.  However, it looks
like we return false because we only check if class files exist in the
list of the Notification's paths if this ContainerClassLoader has the
ability to redefine classes (which is currently disabled).

This fix checks for *.class files before checking whether we can
redefine or not. If there are no *.class files, we return true.  If
there are, then we check whether we can redefine those classes or not.