Change isMinorUpdate behavior to not restart app on JSP changes
This change checks the files being modified and if the file's extension
matches one known to be used for JSPs (i.e. .jsp, .jspx, *.tld, *.tag,
etc.) - and unless there are other files changed that do not match, then
the isMinorUpdate method will return true - indicating that the app
should not be restarted.

Class files are also added to the list of built-in extensions - this is
to enable class redefining which supports a "hot-deploy".  This would
enable users to make minor changes to a class file (i.e. change the
internal body of a method) which would allow the Liberty server to take
advantage of JVM class redefining to make that change without restarting
the app.  This is currently not supported in Liberty, but may be enabled
in the future.