Use StaticValue to multiplex TopicData for static Topic instances
The com.ibm.websphere.event.Topic class is held by many
static final Topic variables throughout the code.  This issue
is that the Topic object holds a reference to an object
of type com.ibm.ws.event.internal.TopicData.  This object
holds state for a specific instance of the EventEngine.
In particular it holds the ExecutorService to post events
to and it holds the list of event handlers.

This type of state needs to be multiplex'ed when loading more
than one Liberty instances with the same set of classes on
a flat class path.