Cache buses based on module
Currently we cache buses per client. This only improves performance if
the same Client makes repeated calls to the same target. This change
allows all Clients in the same module to benefit from targets created by
other Clients in the same module. It also will only clear the cache of
entries from the same module when all Clients in that module are closed
 - so we don't flush the whole cache when we know there are other
Clients who could use the cache.
one client is closed.