Make jsfContainer work for WAR in EAR
When using jsfContainer in an EAR application, it is not sufficient to
apply AppExtLibs to the top level classloader.  Instead, we need to apply
AppExtLibs to the module classloaders.

As a result of this change, we now apply AppExtLibs to the top level
classloader for WAR applications (single module, single classloader) and
for EAR applications we apply them to each individual child module.