Tests for returning Class as a subresource locator
New in the JAX-RS 2.1 spec (section 3.4.1) is the ability for sub
resource locator methods to return a Class type.  According to the spec
the JAX-RS implementation must create a new instance of that class via
a suitable constructor (section 3.1.2 defines what a suitable ctor is),
and then invoke that instance as it would for any other type returned
from the sub resource locator method.