Issue #874: Fix EJB remote delay
The following 3 improvements have been made:

1 - There is no need to wait for the EJB remote runtime
to start if it has already started. The basic bug was
that the EJB Runtime wasn't checking to see if the remote
runtime had already started before decideing to wait for
it to start.

2 - Once any thread has waited for the EJB remote runtime
to start, then there is no point in any threads continuing
to wait, as the EJB remote runtime just isn't going to start;
so eliminate any delay and fail fast.

3 - Small optimization during feture removal; no sense in
doing string comparisons if there is nothing to cleanup.