Put error.badVersion back into kernel.boot's LauncherMessages.nlsprops
This message was removed at some point with a message indicating that
it was only used by com.ibm.ws.kernel.boot.cmdline, but there are at 
least two places (both EnvCheck.java - though in different bundles) that
use this message. It seems that one of them is looking for that message
from this LauncherMessages file.