Prevent "memory leak" in JAXRSClientImpl
The JAXRSClientImpl instance (created when a Client instance is created)
holds a reference to a WebClient object that is part of the WebTarget
for every root-level WebTarget created by that client.  The WebClient is
only stored so that it can be closed appropriately when the Client is
closed.

This change resolves the "leak" (not really a leak since the WebClients
are closed and released when the Client is closed) by making the Set
a WeakHashSet and adding synchronization to avoid race conditions.