Opentracing support for excluding/including spans by URI. (#909)
There are two types of configuration elements: `excludeSpans` and `includeSpans`.

    <opentracing>
      <excludeSpans pattern="URI" type="INCOMING|OUTGOING|BOTH" ignoreCase="true|false" regex="true|false" />
      <includeSpans pattern="URI" type="INCOMING|OUTGOING|BOTH" ignoreCase="true|false" regex="true|false" />
    </opentracing>

These are called filters. Each filter may be specified 0 or more times.

When a JAX-RS request comes in to the container, all `INCOMING` or `BOTH` filters are checked with the incoming URI. If all checks return true, then a new span is created for the incoming request; otherwise, no new span is created, but any incoming spans are passed through.

When a JAX-RS request is sent from the container, all `OUTGOING` or `BOTH` filters are checked with the outgoing URI. If all checks return true, then a new span is created for the outgoing request; otherwise, no new span is created, but any incoming spans are passed through.

The default type of a filter is `BOTH`.

All filters have a `pattern` attribute which is the URI to match to the incoming and/or outgoing URI. If the `pattern` starts with `http(s)://`, then the filter will compare to the [absolute URI](https://docs.oracle.com/javase/9/docs/api/java/net/URI.html#toString--); otherwise, the filter will compare to the URI starting with the [path](https://docs.oracle.com/javase/9/docs/api/java/net/URI.html#getRawPath--) (`/[...]`). In both cases, the filter will compare with [query](https://docs.oracle.com/javase/9/docs/api/java/net/URI.html#getRawQuery--) and [fragment](https://docs.oracle.com/javase/9/docs/api/java/net/URI.html#getRawFragment--), if available, for both the `pattern` and the URI.

If `regex` is false (the default) and `pattern` ends with `*`, then the filter will check if the URI starts with the `pattern`; otherwise, the filter will do an exact match. In both cases, the comparison will be done in a case sensitive or insensitive way depending on the value of `ignoreCase`. The default of `ignoreCase` is `true`.

If `regex` is true, `pattern` is compiled into a [java.util.regex.Pattern](https://docs.oracle.com/javase/9/docs/api/java/util/regex/Pattern.html) and the URI is matched interpreting the `pattern` as a regular expression. The comparison will be done in a case sensitive or [insensitive](https://docs.oracle.com/javase/9/docs/api/java/util/regex/Pattern.html#CASE_INSENSITIVE) way depending on the value of `ignoreCase`. Regular expressions are discouraged because they are resource intensive and run on every JAXRS request. One common mistake when using regular expressions is that some URL components such as the query separator (`?`) have special meaning in regular expressions and must be escaped with `\`.

If no filters match the URI, then the default is that the span is created.

The filters are processed in the order in which they appear in the configuration. For example, if we want to only process URIs that start with `/api`, then we can use an `excludeSpans` first with a `pattern` of `*` to exclude all URIs (this is needed since, by default, spans are created), and then we can use an `includeSpans` with a `pattern` of `/api*`.

Both the `pattern` and URI are compared with [RFC 2396](https://www.ietf.org/rfc/rfc2396.txt)-compliant % URI-encoding, when needed.