[997] Initial Microprofile Rest Client impl
This includes just the basics - creating a new RestClientBuilder that
can produce a type-safe client from an annotated interface.

The spec code is added directly for now, but will eventually be
stored in maven central and artifactory.

Unit tests include finding the appropriate builder, creating a non-
null client, and verifying that the baseUrl must be set prior to
building.