#1479 Don't use the TCCL when choosing a parser impl for beans.xml
This fix ensures that beans.xml is parsed using the SAXParserFactory 
that is visible to Liberty components.  The Weld code loads a parser
factory from the thread context classloader while the app is starting.  
This means that it could load a parser factory from the user application 
which could have undesired consequences.  This fix resolves it by 
temporarily setting the thread context classloader to be the CDI bundle 
loader - this means that Liberty will find the parser factory that ships 
with Liberty or the JDK, rather than any that might ship with the app.

This commit also includes a test case of this scenario where the app 
provides a parser factory that simply throws an exception.  The test
fails if it detects that exception, and passes if the app starts without
exception.