Updates to common JavaScriptUtils class
Enhances the `JavaScriptUtils` class to allow for not HTML-encoding cookie names and values if, for instance, the value needs to be set using pure JavaScript and requires a break in the actual string value itself.

Example:

```JavaScript
document.cookie="name="+encodeURI(location)+";";
```

This also adds public methods to pass in specific cookie properties that should be set.