Use module classloader when creating the bus
Usually the createClientScopedBus method is called with 
the web module's loader set as the current thread's 
context classloader.  However, there are cases where this 
method can be invoked as a result of injection (called via 
the creation of JAX-WS metadata where the thread context
classloader is set to an OSGi bundle's classloader.  As a
result, the LibertyApplicationBusFactory must explicitly 
set the thread context classloader to the module loader 
when creating the bus.