Parameterize JSON4J providers
In some cases, JSON4J-based return types are getting
processed by Jackson instead of the JSON4J providers. This
is because the JSON4J providers are not parameterized.
This commit resolves this by parameterizing the providers
with the JSONArray and JSONObject type so that it will be
sorted higher in the list for providers matching those
types.