Set log mark before deploying CDI extension app
It seems like in some cases the app may be already deployed. The test
framework copies the app to dropins and then looks for the app started
message. Without setting the log mark, it will find the earlier message
from when the old app was started, meaning that it will start making
requests before the app is ready.