Update jsonb_fat to not mix JSONB and JSONP providers
Normally Yasson works with the Glassfish RI and Johnzon-jsonb works with Johnzon-core.  In the JSON-B FAT we have some tests that confirm using non-standard pairs such as Yasson + Johnzon-core works.

Due to Glassfish RI and Johnzon-core interpreting the JSON-P spec differently, this no longer works, so I have opened up an issue on the JSON-P spec to request clarification: https://github.com/eclipse-ee4j/jsonp/issues/78

Until jsonp#78 is resolved, we need to update the tests to not mix providers.