Issue #2255: Cache common fields to improve JSON logging performance
Fields such as host, wlpUserDir, ibm_userDir, serverName, ibm_serverName stay the same for the life of the server and fields such as type stay the same for each event type. Instead of generating the string for these fields everytime for every individual event, the string will be generated the first time this type of event occurs and then stored so that it can be used for all future occurences of this event.