Fix MicroProfile FATs which repeat tests at the class level
When using RepeatTests at the class level (rather than the suite level)
you need to specify which servers should be updated, otherwise your
updates don't get propagated to the server.

In addition, the server needs to be stopped at the point that the server
config is updated with the new feature.