Uninstall leftover app configurations on restart
Must avoid uninstalling application configurations when
stopping an application during server shutdown.  This
is because config admin will not process the removal if
the framework is shutting down.  Instead the config bundle
must remain installed and will be cleaned up on restart.
This way config admin will have a chance to clean out
the application configuraiton before any applications
start.