Direct stderr to the same output file when running Maven in the build
This fixes a problem where _fat_tck projects would hang if they
outputted enough data to stderr. If nothing was consuming stderr, the
surefire thread that was processing messages would hang trying to write
to it and stop processing messages. This would either result in the
whole build hanging, or in surefire reporting that the forked vm did not
shut down properly at the end of the test run.