Remove sync block to improve concurrency in WebContainer
Use putIfAbsent instead of containsKey / put in WebContainer.

Update to call IServletWrapper.addServletReferenceListener outside of
the constructor so that we aren't adding in wrappers we don't actually
keep around and keep the logic consistent.

Make addServletReferenceListener synchronized in order to not have
issues with concurrency when adding to the list.