Repackage Commons Compress as com.ibm.ws.* package
Various kernel boot utilities like server package minify require the
Apache Commons Compress APIs.  However, by putting them in the kernel
boot bundle (which is also used in the classpath to start the Liberty
server) it ends up on the classpath for applications.  This causes
unintended classloading behavior when customer applications attempt to
package their own version of Commons Compress.  In order to ensure that
we can use it, and ensure that user apps can't load it, we need to 
re-package these classes into com.ibm.ws.* packages.