Suggestion for improvement of ENTRY logging
Improves scenarios where otherwise the difference between entry(null) and entry("null")
as well as entry("") and entry() is not easily identifiable.

The new format introduced by this commit is only a discussion proposal, and once
accepted that it makes sense, will also need to be added to other codepaths.