Lazily initialize CacheManager in sessionCache-1.0
Due to a race condition in dynamic the ShareLibrary implementation,
a modification event may propagate to a depending service before the
ClassLoader is updated. We have opened issue #2917 to hopefully get
this fixed.

In the immediate term, we will work around this race condition by lazily
initializing the CacheManager, instead of eagerly initializing it in
CacheStoreService activation.