Expose CDI spec deployment errors in the server log
When an exception is thrown during app deployment, check whether the
cause chain includes an exception defined by the CDI spec. If so, wrap
the thrown exception with a newly created instance of the CDI spec
exception.

This ensures that the CDI spec defined exception is shown in the server
log.