Issue #3226: Fix NullPointerException in EJSContainer when runUnderUOW
When UOWManager.runUnderUOW is being used, the EJBContainer needs
to reset the ContainerTx associated with the current EJB method
call, as runUnderUOW may have switched the transaction that
the EJBContainer believes to be active.

When UserTransaction.begin is called, EJB Container is notified
and records if runUnderUOW is in play; then resets the ContainerTx
during postInvoke.