Issue 3425: Clear the GlobalClassloadingConfiguration service's properties.
During server shutdown, a shutdown hook could trigger a MalformedURLException: unknown protocol: wsjar.
This is due to the OSGi framework having already shutdown and the wsjar protocol handler being unavailble.
The solution was to null out the props for GlobalClassloadingConfiguration, so that wsjar is no longer the
protocol used for zip archives.