Fix leak of WABDeployedAppInfo when a wab is redeployed many times
The issue is wabGroups map never clears out old WABGroup objects
when the WABGroup is empty.  This leaves a stale WABDeployedAppInfo
object in the wabGroups map. The next time the WAB is redeployed
it reuses the stale WABGroup this ends up leaking out the
new WABDeployedAppInfo even when it is not used, but should be.

The fix is to ensure the stale WABGroup is removed so it is not
reused.