Issue 3489: Allow outbound MP Rest Clients to use SSL configuration
This fix allows users to use custom SSL certificates by configuring
the key/trust store in server.xml with their MP Rest Client.  This
functionality has existed with JAX-RS 2.0 Clients for some time, but
did not work for MP Rest Clients.

This requires a new auto-feature that is enabled when mpRestClient-1.0
and ssl-1.0 are enabled.  This auto-feature will pull in the
jaxrsClient-2.0 (tolerating 2.1) feature and will register a listener
for new rest clients - that listener will then add the
LibertyJaxRsClientSSLOutInterceptor which utilizes the SSL config on
outbound calls.

This is tested by modifying the existing basic test case to use a
non-default SSL configuration.