Issue #3604: Fix MessageEndpointProxy with ME method overrides
The EJB specification requires that the MessageEndpoint proxy
handle the MessageEndpiont interface methods and route other
methdods to the message-driven bean implementation. However,
currently if the message-driven bean implements a method with
the same signature as one of the MessageEndpoint methods, then
the generated Proxy fails to load with a ClassFormatError
because it contains 2 methods with the same signature
(in the no-method-interface scenario); otherwise the MessageEndpoint
method is just incorrectly routhed to the bean when called by
the adapter.

To fix this; the MessageEndpoint interface methods are removed
from the list of EJB method infos; thus the code that generates
the proxy classes will not attempt to add them as EJB methods,
and will only provide the requried MessageEndpoint implementations.