Rewrite repository resolver to use kernel resolver
The new repository resolver delegates to the kernel resolver to resolve
features. This allows it to use the existing logic for resolving feature
dependencies using tolerates.

This fixes a problem where we couldn't resolve features in openliberty
because there are a few cases where a feature has a direct dependency on
a feature which does not exist in openliberty, but tolerates a later
version of that feature which does exist in openliberty.

Broad overview of the changes:

KernelResolverEsa and KernelResolverRequirement translate an
EsaResource into a ProductFeatureDefinition which is understood by the
kernel resolver

KernelResolverRepository implements FeatureResolver.Repository to
provides an interface to features in the repository.

RepositoryResolver has been completely rewritten to use the kernel
resolver with the support classes listed above.

Removed all code related to the old resolver

Added unit tests for integration with the kernel resolver, and for the
RepositoryResolver class.

Converted existing integration tests to unit tests to ensure correctness
and compatibility with the old resolver implementation.

Added a separate method to do resolution as a set. This isn't used yet
but should allow better resolution when installing based on a
user-provided server.xml.

Remove support for resolving ifixes as we will no longer support
installing ifixes from the repository This should not affect anyone
since we haven't used this feature since around 8.5.5.2.