Issue 1425: Suppress CWWKW1001W and CWWKW1002W msgs in JAX-RS 2.1
These messages indicate a mismatch between the lifecycle scopes of
JAX-RS and CDI for resources and providers, respectively.  In most cases
this mismatch is not worth logging since a user is indicating their
preference for the particular scope by adding a CDI or EJB annotation
to their resource/provider class.

The message will still be output, but only when tracing is enabled.
Trace full message since Tr.debug(...) does not translate.

The CDITest verifies that this message is suppressed by no longer
allowing that message to be ignored - if the message appears during that
test, then it will fail.