Enable users to configure KeepAlive setting in JAX-RS Clients
A user can now enable (default) or disable the HTTP Keep-Alive setting
by setting the property com.ibm.ws.jaxrs.client.keepalive.connection to
either "keep-alive" (default) or "close" respectively.

At this point, there is not much value in enabling KeepAlive as we do
not currently re-use connections.

The test ensures that the Connection header is affected by the setting
of this property.