Improve performance of adding providers to ProviderFactories.
Change from using CopyOnWriteArrayList to using an AtomicReference and
creating a new List each time we need to add new Providers to the list.
Also update the 2.1 CXF JAXRS code to use the
ProviderFactory.getProviderProduceTypes method like is done in the 2.0
JAXRS code.
With the changes to not use CopyOnWriteArrayList, now the concurrent
issues that were resolved only with Java 8 will not also be resolved
with Java 7.