JWT SSO FAT: Update JWK test to use different keystore
Ensure that the JWK test for the `jwtSso-1.0` feature sets the `mpJwt` configuration to use a keystore with a certificate that can't be used to validate the JWT cookie, but also configures the `jwksUri` attribute. This should make sure we're testing that if the `jwksUri` attribute is configured, we should use the JWK endpoint to get the right key to verify the JWT cookie instead of looking in the configured keystore.