Fix default ID for FeatureReplacementAction
Previously, the ID was initialized on creation, before the object knows
which features it's adding and removing. This results in an ID string
of "FeatureReplacementAction REMOVE null ADD null".

With this change, unless an ID is specified manually, the ID is computed
dynamically. This should be safe as the features to add and removed are
set before the test runs and not modified.