Only check if FT annotations are disabled
Fault Tolerance provides a way of disabling the fault tolerance logic
which would be added by an annotation. isAnnotationEnabled checks
whether a specific FT annotation should be ignored.

However, we shouldn't be taking any action for non-fault tolerance
annotations based on the results of this method, so make it an
exception to pass a non-FT annotation to this method.