Web container performance improvements.
Update query string parsing again to remove some additional overhead.
This optimizes the cases that a key and/or value does not have a % or a
+ character which is likely the normal case.  The optimization also
removes creating char[]'s for the normal case.
Also update the SRTInputStream31 class to not use Boolean's, but instead
to use booleans.  There is no need for a null state in the code so the
Boolean being the type is unnecessary.