Issue 4339: Ensure GW bundle is started before allowing classloads
This root problem of this issue appears to be that the gateway bundle
is not activated until after the application has attempted to use it's
classloader to load classes. This fix resolves the issue by preventing
the createGatewayBundleClassLoader method from exiting until the bundle
is started or a timeout (default 10 seconds) has expired.