Issue #4465: Fix EJB Timer Create Timing Issue
This error may occur when creating a persistent Calendar based EJB timer, which only has one
scheduled time, and that time expires while the timer is being created.

Specifically, the EJB Container checked the scheduled expression to ensure it has at least
one expiration in the future. If there is at least one timeout, then the timer is created
and scheduled to run. When scheduled, the first timeout is calculated again. However, if
the one timeout for the timer has passed between the time the EJB Container checked and
the timer is scheduled, then the RejectedExcecutionException occurs and timer creation fails.

The EJB specification indicates an EJB timer should determine timeouts based on the time
the EJB timer is created, not when the transaction commits, so the timer should be valid
and should run immediately.

The fix will be to detect the small timing window, and schedule to run the timer immediately,
avoiding the failed timer creation.