Update the Fault Tolerance CDI integration for 2.0
Allow CompletionStage return type from interceptor

Abstract out the creation of AnnotationConfig to a factory so we can
provide a different implementaion for 2.0.

Change the AggregatedFTPolicy to use the new buildAsync method.

Create new project and feature for 2.0 with new implementation

Use equality check for async method return types

* Assert that an async method must return a specific type, not just a type
which assignable to that type.

* This is important because we have to provide a proxy object back to the
user as a result, and this is made much harder if they can declare that
a subtype is returned.

Add message to allow CompletionStage return type

* CompletionStage is now allowed as the return type from an asynchronous
method. Update the error message accordingly.

Remove exception for private async methods

* Async methods must return Future or CompletionStage, this includes
private methods.