MP Rest Client CompletionStage completing for subsequent stages
This fixes an issue where the CompletionStage returned by instances of
MicroProfile Rest Client interfaces do not execute subsequent stages.
The method completes normally and the remote invocation occurred
asynchronously, but if the user attempted a `then*` type method on the
returned `CompletionStage`, that stage would not execute because the
internal implementation assumed that the initial stage was not complete.
This fix resolves that issue.