Ignore some FFDCs in Fault Tolerance Async
Looking at the history, I just forgot to add these FFDCIgnores when
I added some catch blocks.

In each case, the catch would be triggered if the user's method threw an
exception. As Fault Tolerance is concerned with handling failures, this
isn't unexpected so it doesn't make sense to FFDC here.