Fix for reporting false positives for conflicts
Previously the feature resolver would give up once
it found a hard conflict with no optional candidates
to choose from.  It would stop pushing alternative
paths to take to try and find a solution with less
conflicts.  It would just take the first solution
tried which will contain the hard conflict, but
may also contain other conflicts that are possible
to avoid if other solutions were tried.

This fix now tries to find a better solution that
has as few conflicts as possible.  This helps
weed out the false positives.  It is still possible
to get some confusing results if the problem set
includes choices that result in the same number of
conflicts but with different root causes.  The
resolver simply picks one of the solutions with
the least conflicts and reports the conflicts for
that solution.