Remove liberty change flags for function ported to the cxf community.
- Remove liberty change flags for changes that went into apache/cxf#407
and that are now in CXF 3.1.16 that was recently brought into open
liberty. This also pulls in the community way of doing one of the
changes that is different than what IBM originally proposed.
- Also update ProviderFactory to use the commonBaseClass as part of the
cache key like is done with the 3.2 code now that that the 3.1 code uses
a commonBase Class.