Allow CXF-specific client properties for JAX-RS Client
With this change, users can specify CXF-specific client properties when
using the jaxrs-2.0 or jaxrs-2.1 feature.  These properties should be
specified on the Client or ClientBuilder via the property(...) method,
and the property and value should be specified as a String (i.e. 
`client.property("client.ChunkingThreshold", "8000")` ).

The property name should be prefixed with "client." - i.e. 
"client.AllowChunking", "client.ConnectionTimeout", etc.