Add deployment options to ShrinkHelper
Allows the deployer to skip writing to the publish/servers directory and
to force an overwrite when deploying to a liberty server or client.

Skipping the server directory is useful when a test deploys and removes
several apps (e.g. when testing deployment)

Overwriting is useful when using ShrinkHelper in conjunction with
RepeatTests at the class level and a slightly different app is deployed
on each test repetition.