Fix to preserve order when adding chains for postponed decisions
The chains are sorted to rank the ones with lower preferred
versions first, but the chains with equal preferred versions end
up getting sorted randomly with each other.  This ends up randomly
blaming one of the chains when a conflict occurs. This fix now
ensures the chains with equal preferred version are inserted by
insertion order so that the first ones added will be blamed first.

This is important for reporting conflicts to ensure earlier
decisions are blamed first. Otherwise we can get confusing error
messages about the root features in conflict.

This commit also fixes an issue where when a hard selection must
be made that causes a conflict in existing postponed decisions.
We must add the conflict from the postponed decision but we must
also continue to select the hard choice that introduces the
conflict.