Pull in 3.2.4 version of SseBroadcasterImpl
This avoids a change made in 3.2.5 or 3.2.6 that breaks Liberty's impl.
The change forces users to use the CXF SseEventSinkImpl rather than a
custom impl (like our LibertySseEventSinkImpl).  At some point, we
should start to use the CXF impl, but for now, we'll use the 3.2.4
version of SseBroadcasterImpl so as to continue working with our
existing LibertySseEventSinkImpl.