Avoid TCKs creating a top level failure
Avoid creating a test failure when mvn returns with a non-zero return
code, as long as there are tests that failed.

Usually, a non-zero return code means that a test failed. In this case,
we don't want to create an additional test failure since we will report
the actual failures and having an extra failure confuses the build
monitoring tooling.