Update Metrics and OpenAPI TCKs to use common code
MvnUtils has been updated to allow custom properties to be set on the
mvn command line. This should allow Metrics and OpenAPI to use the
commond code path for launching their TCKs and processing the results.

This also means they'll pick up the new change to avoid creating a
top-level failure if any of the tests fail.