Issue #5136: Avoid Loading CORBA classes if remote EJB is not enabled
There is no need to generate and/or load the Tie classes for remote
EJB interfaces if the ejbRemote feature is not enabled. This is
important for Java 11, which doesn't provide the CORBA classes
used by the Ties.

The remote wrappers are still generated/loaded to allow an appropriate
error message if an attempt is made to use them when the ejbRemote
feature is not enabled.

Similar to issue 5011, but now handles EJB 2.x home ties.