Fix EnablingBeansXmlValidationTest
Remove dependency on exact WELD exception message for 2.4.7

Previously we looked for a validation exception, now we're getting a
parse error, but really we don't mind exactly how the error manifests.

Weld 3 appears to use a different code path and reliably produces the
same error message so I haven't changed that test.