Update test case to behave like other tests in the suite.
I have hit the testAutoExtractWarDropins() test failing in 3 personal
builds.  The change to the test is to ignore a warning that is possible
when stopping the server for the type of scenarios done by this test.
The other tests in the suite already have this logic.  Only this one was
missing it.