Make Asynchronous methods return when they timeout
This change adds the ability to abort the execution of an asynchronous
method and set its result immediately.

We can then use this to abort execution of an asynchronous method when a
timeout occurs.

However, we can't do this if Retry or Fallback are in use as they need
to capture and handle the exception and return a different result.

In addition, some refactoring of ExecutionContextImpl and QueuedFuture
was required to ensure that we take the correct code paths when starting
execution of an asynchronous method. The exising code could cause the
calling thread to be interrupted if the timeout occurred before the
asynchronous method started running.