Issue 5439: Use new DateFormat for each request
The DateFormat is not thread safe, so using the same one for multiple
requests risks issues like ArrayIndexOutOfBoundsExceptions.

This change removes the Headers.java file, as the single static usage
of SimpleDateFormat was the only unique change from what gets pulled in
from the Apache CXF community. All of the other changes had been
previously contributed back to the open source community.