Issue #5585: Fix EJB ScheduleExpression serialization
Provide a custom implementation of ScheduleExpression that handles the
serialization incompatibility between the RI implementation for EJB 3.2 and
the version used for EJB 3.1. The serial version identifier is the same for
both implementations, however the instance variables have different names,
so default serialization is not able to restore the values when deserializing
bytes from the older implementation using the RI implementation.

Since an instance of ScheduleExpression is serialized and stored in a
database for persistent calendar based EJB timers, supporting the
deserialization of both versions of the class is needed for applications
migrating from EJB 3.1 to EJB 3.2.