Update to convert integer to String in order to write the right thing.
If you pass an int to a write method it treats it as a character and not
an integer.  In order to write it you need to confirm the integer to a
String so that it is the right character number.  This caused us to
think we were always running with a different JDK for each server
restart.