Update Fault Tolerance FATs to run on FT 2.0
All tests run on FT 2.0 and 1.1 because the implementation for these two
versions is very different so, in most cases, testing on one does not
give confidence that the other will also pass.

A few tests also run on FT 1.0, so that we have some coverage there.
Most of the code for 1.0 is shared with 1.1.