Fix case where timeout stop called before start
Previously, the AsyncExecutor could not start the timeout until after it
had submitted the task, because the timeout callback requires the
ExecutionReference for the running task.

However, this meant that in rare cases the task could run, finish and
timeout.stop() could be called before timeout.start() was called.

To prevent this, we split the starting of the timeout and the setting of
the timeout callback. This allows the timeout to be started before
submitting the task and the callback to be set afterwards.

Also took the opportunity to tighten up the state handling logic in
Timeout to help ensure that any more incorrect handling of timeout
states will be caught.